# huff0: translate asm implementation into avo program