Currently register allocation fails for some reason. Fixes #529