@klauspost 2022-03-24
Some notes:

* Don't pre-allocate and re-use temp vars. Instead "allocate" them just as you need them. In your code, you keep "temp" allocated across calls since you pass it as a parameter. This gives the allocator more space to work.
* Dereference the bitreader between calls, so you don't need to keep it alive all the time.
* Made bmi/non-bmi versions. We already have the detection code and don't require the cleanup.
* Moved "off" to memory. Used the return index which is already allocated, which gets updated inplace.
* Moved RCX/CL references to a variable to make it more clear what is happening.

The `8H` is limited to just a few registers (which you probably know), which avo needs to know up-front. Maybe there is a bug in avo. I used manual assignment for those.

"Compiling" version: https://gist.github.com/klauspost/8f8dbbd9745662464dfac37d00cbd5f6