@WojciechMula 2022-03-25
There are significant regressions. The asm code cached more values in GPRs.

```
benchmark                                                   old ns/op     new ns/op     delta
BenchmarkDecompress4XNoTable/digits/100-16                  424           421           -0.59%
BenchmarkDecompress4XNoTable/digits/10000-16                11797         53166         +350.67%
BenchmarkDecompress4XNoTable/digits/262143-16               316076        1393336       +340.82%
BenchmarkDecompress4XNoTable/gettysburg/100-16              326           322           -1.23%
BenchmarkDecompress4XNoTable/gettysburg/10000-16            11727         54531         +365.00%
BenchmarkDecompress4XNoTable/gettysburg/262143-16           328128        1462261       +345.64%
BenchmarkDecompress4XNoTable/twain/100-16                   416           413           -0.86%
BenchmarkDecompress4XNoTable/twain/10000-16                 11762         56049         +376.53%
BenchmarkDecompress4XNoTable/twain/262143-16                399616        1521951       +280.85%
BenchmarkDecompress4XNoTable/low-ent.10k/100-16             445           437           -1.91%
BenchmarkDecompress4XNoTable/low-ent.10k/10000-16           11643         44926         +285.86%
BenchmarkDecompress4XNoTable/low-ent.10k/262143-16          273885        1175472       +329.18%
BenchmarkDecompress4XNoTable/superlow-ent-10k/262143-16     274214        1176740       +329.13%
BenchmarkDecompress4XNoTable/case1/100-16                   410           409           -0.39%
BenchmarkDecompress4XNoTable/case1/10000-16                 11863         47098         +297.02%
BenchmarkDecompress4XNoTable/case1/262143-16                298821        1226254       +310.36%
BenchmarkDecompress4XNoTable/case2/100-16                   421           416           -1.33%
BenchmarkDecompress4XNoTable/case2/10000-16                 11657         48478         +315.87%
BenchmarkDecompress4XNoTable/case2/262143-16                291177        1272254       +336.93%
BenchmarkDecompress4XNoTable/case3/100-16                   408           407           -0.17%
BenchmarkDecompress4XNoTable/case3/10000-16                 11664         49914         +327.93%
BenchmarkDecompress4XNoTable/case3/262143-16                294305        1297013       +340.70%
BenchmarkDecompress4XNoTable/pngdata.001/100-16             439           429           -2.23%
BenchmarkDecompress4XNoTable/pngdata.001/10000-16           11965         45253         +278.21%
BenchmarkDecompress4XNoTable/pngdata.001/262143-16          307526        1215161       +295.14%
BenchmarkDecompress4XNoTable/normcount2/100-16              337           329           -2.17%
BenchmarkDecompress4XNoTable/normcount2/10000-16            11978         50669         +323.02%
BenchmarkDecompress4XNoTable/normcount2/262143-16           303912        1330331       +337.74%
BenchmarkDecompress4XNoTableTableLog8/digits-16             113774        531650        +367.29%
BenchmarkDecompress4XTable/digits-16                        114840        532410        +363.61%
BenchmarkDecompress4XTable/gettysburg-16                    3447          9315          +170.23%
BenchmarkDecompress4XTable/twain-16                         402565        1520643       +277.74%
BenchmarkDecompress4XTable/low-ent.10k-16                   43690         179044        +309.81%
BenchmarkDecompress4XTable/superlow-ent-10k-16              12669         47572         +275.50%
BenchmarkDecompress4XTable/case1-16                         2078          2102          +1.15%
BenchmarkDecompress4XTable/case2-16                         2037          2064          +1.33%
BenchmarkDecompress4XTable/case3-16                         2036          2040          +0.20%
BenchmarkDecompress4XTable/pngdata.001-16                   58369         236666        +305.47%
BenchmarkDecompress4XTable/normcount2-16                    1437          1432          -0.35%

benchmark                                                   old MB/s     new MB/s     speedup
BenchmarkDecompress4XNoTable/digits/100-16                  236.15       237.51       1.01x
BenchmarkDecompress4XNoTable/digits/10000-16                847.70       188.09       0.22x
BenchmarkDecompress4XNoTable/digits/262143-16               829.37       188.14       0.23x
BenchmarkDecompress4XNoTable/gettysburg/100-16              307.23       311.03       1.01x
BenchmarkDecompress4XNoTable/gettysburg/10000-16            852.73       183.38       0.22x
BenchmarkDecompress4XNoTable/gettysburg/262143-16           798.90       179.27       0.22x
BenchmarkDecompress4XNoTable/twain/100-16                   240.16       242.22       1.01x
BenchmarkDecompress4XNoTable/twain/10000-16                 850.20       178.41       0.21x
BenchmarkDecompress4XNoTable/twain/262143-16                655.99       172.24       0.26x
BenchmarkDecompress4XNoTable/low-ent.10k/100-16             224.64       228.97       1.02x
BenchmarkDecompress4XNoTable/low-ent.10k/10000-16           858.90       222.59       0.26x
BenchmarkDecompress4XNoTable/low-ent.10k/262143-16          957.13       223.01       0.23x
BenchmarkDecompress4XNoTable/superlow-ent-10k/262143-16     955.98       222.77       0.23x
BenchmarkDecompress4XNoTable/case1/100-16                   243.67       244.60       1.00x
BenchmarkDecompress4XNoTable/case1/10000-16                 842.98       212.32       0.25x
BenchmarkDecompress4XNoTable/case1/262143-16                877.26       213.78       0.24x
BenchmarkDecompress4XNoTable/case2/100-16                   237.31       240.51       1.01x
BenchmarkDecompress4XNoTable/case2/10000-16                 857.84       206.28       0.24x
BenchmarkDecompress4XNoTable/case2/262143-16                900.29       206.05       0.23x
BenchmarkDecompress4XNoTable/case3/100-16                   245.01       245.47       1.00x
BenchmarkDecompress4XNoTable/case3/10000-16                 857.37       200.35       0.23x
BenchmarkDecompress4XNoTable/case3/262143-16                890.72       202.11       0.23x
BenchmarkDecompress4XNoTable/pngdata.001/100-16             227.91       233.09       1.02x
BenchmarkDecompress4XNoTable/pngdata.001/10000-16           835.80       220.98       0.26x
BenchmarkDecompress4XNoTable/pngdata.001/262143-16          852.42       215.73       0.25x
BenchmarkDecompress4XNoTable/normcount2/100-16              297.01       303.54       1.02x
BenchmarkDecompress4XNoTable/normcount2/10000-16            834.86       197.36       0.24x
BenchmarkDecompress4XNoTable/normcount2/262143-16           862.56       197.05       0.23x
BenchmarkDecompress4XNoTableTableLog8/digits-16             878.96       188.10       0.21x
BenchmarkDecompress4XTable/digits-16                        870.80       187.83       0.22x
BenchmarkDecompress4XTable/gettysburg-16                    449.10       166.18       0.37x
BenchmarkDecompress4XTable/twain-16                         651.18       172.39       0.26x
BenchmarkDecompress4XTable/low-ent.10k-16                   915.55       223.41       0.24x
BenchmarkDecompress4XTable/superlow-ent-10k-16              828.81       220.72       0.27x
BenchmarkDecompress4XTable/case1-16                         26.46        26.17        0.99x
BenchmarkDecompress4XTable/case2-16                         22.09        21.80        0.99x
BenchmarkDecompress4XTable/case3-16                         23.57        23.53        1.00x
BenchmarkDecompress4XTable/pngdata.001-16                   877.18       216.34       0.25x
BenchmarkDecompress4XTable/normcount2-16                    60.54        60.75        1.00x
```