@WojciechMula 2022-03-31
I managed to fix obvious mistakes, but still, there are regressions. I will investigate it further, just dumping the current state.

```
benchmark                                                   old ns/op     new ns/op     delta
BenchmarkDecompress4XNoTable/digits/100-16                  424           418           -1.20%
BenchmarkDecompress4XNoTable/digits/10000-16                11797         19903         +68.71%
BenchmarkDecompress4XNoTable/digits/262143-16               316076        509543        +61.21%
BenchmarkDecompress4XNoTable/gettysburg/100-16              326           321           -1.32%
BenchmarkDecompress4XNoTable/gettysburg/10000-16            11727         15252         +30.06%
BenchmarkDecompress4XNoTable/gettysburg/262143-16           328128        404332        +23.22%
BenchmarkDecompress4XNoTable/twain/100-16                   416           413           -0.91%
BenchmarkDecompress4XNoTable/twain/10000-16                 11762         15229         +29.48%
BenchmarkDecompress4XNoTable/twain/262143-16                399616        454245        +13.67%
BenchmarkDecompress4XNoTable/low-ent.10k/100-16             445           437           -1.84%
BenchmarkDecompress4XNoTable/low-ent.10k/10000-16           11643         20342         +74.71%
BenchmarkDecompress4XNoTable/low-ent.10k/262143-16          273885        510006        +86.21%
BenchmarkDecompress4XNoTable/superlow-ent-10k/262143-16     274214        509801        +85.91%
BenchmarkDecompress4XNoTable/case1/100-16                   410           408           -0.63%
BenchmarkDecompress4XNoTable/case1/10000-16                 11863         19904         +67.78%
BenchmarkDecompress4XNoTable/case1/262143-16                298821        511902        +71.31%
BenchmarkDecompress4XNoTable/case2/100-16                   421           415           -1.59%
BenchmarkDecompress4XNoTable/case2/10000-16                 11657         19992         +71.50%
BenchmarkDecompress4XNoTable/case2/262143-16                291177        511883        +75.80%
BenchmarkDecompress4XNoTable/case3/100-16                   408           408           -0.07%
BenchmarkDecompress4XNoTable/case3/10000-16                 11664         19853         +70.21%
BenchmarkDecompress4XNoTable/case3/262143-16                294305        510981        +73.62%
BenchmarkDecompress4XNoTable/pngdata.001/100-16             439           428           -2.53%
BenchmarkDecompress4XNoTable/pngdata.001/10000-16           11965         15898         +32.87%
BenchmarkDecompress4XNoTable/pngdata.001/262143-16          307526        404434        +31.51%
BenchmarkDecompress4XNoTable/normcount2/100-16              337           330           -2.14%
BenchmarkDecompress4XNoTable/normcount2/10000-16            11978         19868         +65.87%
BenchmarkDecompress4XNoTable/normcount2/262143-16           303912        511161        +68.19%
BenchmarkDecompress4XNoTableTableLog8/digits-16             113774        195346        +71.70%
BenchmarkDecompress4XTable/digits-16                        114840        196381        +71.00%
BenchmarkDecompress4XTable/gettysburg-16                    3447          3941          +14.33%
BenchmarkDecompress4XTable/twain-16                         402565        457229        +13.58%
BenchmarkDecompress4XTable/low-ent.10k-16                   43690         79427         +81.80%
BenchmarkDecompress4XTable/superlow-ent-10k-16              12669         21837         +72.37%
BenchmarkDecompress4XTable/case1-16                         2078          2083          +0.24%
BenchmarkDecompress4XTable/case2-16                         2037          2026          -0.54%
BenchmarkDecompress4XTable/case3-16                         2036          2024          -0.59%
BenchmarkDecompress4XTable/pngdata.001-16                   58369         79526         +36.25%
BenchmarkDecompress4XTable/normcount2-16                    1437          1432          -0.35%
```