@klauspost 2022-05-02
Running this branch, I get these numbers on `4XNoTable`:
```
λ benchcmp before.txt after.txt                                                               
benchmark                                                   old ns/op     new ns/op     delta 
BenchmarkDecompress4XNoTable/digits/100-32                  334           334           +0.21%
BenchmarkDecompress4XNoTable/digits/10000-32                10835         11017         +1.68%
BenchmarkDecompress4XNoTable/digits/262143-32               303585        310422        +2.25%
BenchmarkDecompress4XNoTable/gettysburg/100-32              285           285           +0.04%
BenchmarkDecompress4XNoTable/gettysburg/10000-32            11393         11480         +0.76%
BenchmarkDecompress4XNoTable/gettysburg/262143-32           327973        333221        +1.60%
BenchmarkDecompress4XNoTable/twain/100-32                   331           332           +0.36%
BenchmarkDecompress4XNoTable/twain/10000-32                 11458         11453         -0.04%
BenchmarkDecompress4XNoTable/twain/262143-32                374970        386719        +3.13%
BenchmarkDecompress4XNoTable/low-ent.10k/100-32             367           372           +1.17%
BenchmarkDecompress4XNoTable/low-ent.10k/10000-32           10812         10956         +1.33%
BenchmarkDecompress4XNoTable/low-ent.10k/262143-32          256684        260314        +1.41%
BenchmarkDecompress4XNoTable/superlow-ent-10k/262143-32     256839        261779        +1.92%
BenchmarkDecompress4XNoTable/case1/100-32                   318           320           +0.72%
BenchmarkDecompress4XNoTable/case1/10000-32                 10803         11021         +2.02%
BenchmarkDecompress4XNoTable/case1/262143-32                277377        280121        +0.99%
BenchmarkDecompress4XNoTable/case2/100-32                   345           342           -1.10%
BenchmarkDecompress4XNoTable/case2/10000-32                 10659         10870         +1.98%
BenchmarkDecompress4XNoTable/case2/262143-32                268723        274144        +2.02%
BenchmarkDecompress4XNoTable/case3/100-32                   333           333           +0.15%
BenchmarkDecompress4XNoTable/case3/10000-32                 10737         10806         +0.64%
BenchmarkDecompress4XNoTable/case3/262143-32                272268        276636        +1.60%
BenchmarkDecompress4XNoTable/pngdata.001/100-32             361           361           -0.11%
BenchmarkDecompress4XNoTable/pngdata.001/10000-32           11583         11683         +0.86%
BenchmarkDecompress4XNoTable/pngdata.001/262143-32          306257        313848        +2.48%
BenchmarkDecompress4XNoTable/normcount2/100-32              287           288           +0.49%
BenchmarkDecompress4XNoTable/normcount2/10000-32            10832         11073         +2.22%
BenchmarkDecompress4XNoTable/normcount2/262143-32           279908        283876        +1.42%
BenchmarkDecompress4XNoTableTableLog8/digits-32             107990        109990        +1.85%
```